[core-rest-pipeline] expose tokenCycler
So that other libraries can re-use its auto-refreshing mechanism.
