+++ common/config/rush/common-versions.json
@@ -64,8 +64,8 @@
     // @azure/test-utils-perfstress should depend on lowest version of @azure/core-http for maximum compatibility, allowing test
     // projects to choose a higher version if desired.
     "@azure/core-http": ["^1.0.0"],
-    // @azure/container-registry is using the beta.1 version for CAE support
-    "@azure/core-rest-pipeline": ["1.1.0-beta.1"],
+    // @azure/container-registry is using the beta.2 version for CAE support
+    "@azure/core-rest-pipeline": ["1.1.0-beta.2"],
     // @azure/event-processor-host is on a much lower major version
     "@azure/ms-rest-nodeauth": ["^0.9.2"],
     // Idenity is moving from v1 to v2. Moving all packages to v2 is going to take a bit of time, in the mean time we could use v2 on the perf-identity tests.
