+++ sdk/confidentialledger/confidential-ledger-rest/package.json
@@ -86,7 +86,7 @@
   "dependencies": {
     "@azure/core-auth": "^1.3.0",
     "@azure-rest/core-client": "1.0.0-beta.2",
-    "@azure/core-rest-pipeline": "1.1.0-beta.1",
+    "@azure/core-rest-pipeline": "1.1.0-beta.2",
     "@azure/logger": "^1.0.0",
     "tslib": "^2.0.0"
   },
