+++ sdk/containerregistry/container-registry/package.json
@@ -76,7 +76,7 @@
   "dependencies": {
     "@azure/core-auth": "^1.3.0",
     "@azure/core-client": "^1.0.0",
-    "@azure/core-rest-pipeline": "1.1.0-beta.1",
+    "@azure/core-rest-pipeline": "1.1.0-beta.2",
     "@azure/core-paging": "^1.1.1",
     "@azure/core-tracing": "1.0.0-preview.11",
     "@azure/core-util": "^1.0.0-beta.1",
