+++ sdk/core-rest/core-client/package.json
@@ -70,7 +70,7 @@
   "prettier": "@azure/eslint-plugin-azure-sdk/prettier.json",
   "dependencies": {
     "@azure/core-auth": "^1.3.0",
-    "@azure/core-rest-pipeline": "1.1.0-beta.1",
+    "@azure/core-rest-pipeline": "1.1.0-beta.2",
     "tslib": "^2.0.0"
   },
   "devDependencies": {
