+++ sdk/core/core-rest-pipeline/src/util/tokenCycler.ts
@@ -10,17 +10,17 @@ import { delay } from "./helpers";
  *
  * @param options - the options to pass to the underlying token provider
  */
-export type AccessTokenGetter = (
+export type AccessTokenGetter<T extends GetTokenOptions> = (
   scopes: string | string[],
-  options: GetTokenOptions
+  options: T
 ) => Promise<AccessToken>;
 
 /**
  * The response of the
  */
-export interface AccessTokenRefresher {
+export interface AccessTokenRefresher<T extends GetTokenOptions> {
   cachedToken: AccessToken | null;
-  getToken: AccessTokenGetter;
+  getToken: AccessTokenGetter<T>;
 }
 
 export interface TokenCyclerOptions {
@@ -112,10 +112,10 @@ async function beginRefresh(
  *
  * @returns - a function that reliably produces a valid access token
  */
-export function createTokenCycler(
+export function createTokenCycler<T extends GetTokenOptions>(
   credential: TokenCredential,
   tokenCyclerOptions?: Partial<TokenCyclerOptions>
-): AccessTokenRefresher {
+): AccessTokenRefresher<T> {
   let refreshWorker: Promise<AccessToken> | null = null;
   let token: AccessToken | null = null;
 
@@ -160,10 +160,7 @@ export function createTokenCycler(
    * Starts a refresh job or returns the existing job if one is already
    * running.
    */
-  function refresh(
-    scopes: string | string[],
-    getTokenOptions: GetTokenOptions
-  ): Promise<AccessToken> {
+  function refresh(scopes: string | string[], getTokenOptions: T): Promise<AccessToken> {
     if (!cycler.isRefreshing) {
       // We bind `scopes` here to avoid passing it around a lot
       const tryGetAccessToken = (): Promise<AccessToken | null> =>
@@ -199,10 +196,7 @@ export function createTokenCycler(
     get cachedToken(): AccessToken | null {
       return token;
     },
-    getToken: async (
-      scopes: string | string[],
-      tokenOptions: GetTokenOptions
-    ): Promise<AccessToken> => {
+    getToken: async (scopes: string | string[], tokenOptions: T): Promise<AccessToken> => {
       //
       // Simple rules:
       // - If we MUST refresh, then return the refresh task, blocking
