+++ sdk/core/perf-tests/core-rest-pipeline/package.json
@@ -7,7 +7,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@azure/core-rest-pipeline": "1.1.0-beta.1",
+    "@azure/core-rest-pipeline": "1.1.0-beta.2",
     "@azure/core-auth": "^1.3.0",
     "@azure/test-utils-perfstress": "^1.0.0",
     "dotenv": "^8.2.0"
